Refactor reset and verify password forms
